analytics/requirements.txt
dbt
dbt-athena-community

aws-cdk-lib
constructs
boto3
